EV-Hybrid-Advanced/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
python-jose
streamlit
requests
python-multipart
pandas
